Update OpenSSH to 9.9p1
Install sshd-session binary

With 9.8p1, sshd was split into a listener (sshd)
and session (sshd-session) binary. Obviously we
need both for a working ssh server.

Fake privilege separation instead of disabling it

OpenSSH has deprecated disabling privilege separation
for quite some time. While they removed the config
option, most of the actual code for this was still there.
So our patch set restored the "UsePrivilegeSeparation"
option and hooked it up to the old code.
Most of that code is gone since 9.8p1, so continuing
maintaining this method would be very difficult now.

Instead, we now keep privilege separation on but use
root and / as user and chroot directory.
This stays with supported code (and thus shouldn't be
difficult to maintain) but still effectively doesn't
have privilege separation.

Unfortunately, we don't have an sshd user on Android
for actual privelege separation.